+++ Classes/Domain/Repository/FaqitemRepository.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Wacon\FaqSeo\Domain\Repository;
+
+use TYPO3\CMS\Extbase\Persistence\Repository;
+
+class FaqitemRepository extends Repository
+{}